+++ src/refactoring/components/AdminPage.tsx
@@ -8,7 +8,7 @@ export const AdminPage = () => {
       <Layout.Title>관리자 페이지</Layout.Title>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
         <ProductPanel />
-        <CouponPanel  />
+        <CouponPanel />
       </div>
     </Layout.Container>
   );
