+++ src/refactoring/components/admin/product/addProduct/index.ts
@@ -1,0 +1,3 @@
+import AddProductSection from "./AddProductSection";
+
+export { AddProductSection };
