+++ src/refactoring/components/admin/product/editDiscount/DiscountInfo.tsx
