+++ src/refactoring/components/admin/product/editDiscount/EditDiscountSection.tsx
