+++ src/refactoring/components/admin/product/editDiscount/index.ts
@@ -1,0 +1,3 @@
+import EditDiscountSection from "./EditDiscountSection";
+
+export { EditDiscountSection };
