+++ src/refactoring/components/admin/product/editProduct/index.ts
@@ -1,0 +1,3 @@
+import EditProductSection from "./EditProductSection";
+
+export { EditProductSection };
